face-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
torch==2.2.2
torchvision==0.17.2
facenet-pytorch==2.6.0
numpy==1.26.4
pillow==10.2.0
requests==2.32.3
python-dotenv==1.0.1
